Add CMake support
Can do the same as the Makefile now, except the tests, which I never
actually developed very well anyways. It was just a test to see whether
I can test the cli application like this.

CMake is supposed to replace the Makefile soon.

Also telling gcc to show more warnings.