Handle subcommand parsing better
Now when we just type `roshi` it prints it's usage, and doesnt return an
error code.

When we type something as the second parameter, which is supposed to be
the workoutlog/db, it prints that such a file does not exist if it
doesnt.

When we type a correct db but a unknown subcommand we print this too.