Add error handling to config parsing
Config errors out now if we have options in it that are not defined.
Makes it easier for users to find typos.

And we stop execution once there was a parsing error.

We also save in case no config is used. Might be that we want to display
this later.